fix(rename): renamed RDP to External
