+++ .github/workflows/development_deploy.yml
@@ -434,10 +434,10 @@ jobs:
         inlineScript: |
           az webapp restart --name tfs-${{ env.environment }}-${{ env.repository }}
 
-  #################### 7. RDP API ####################
+  #################### 7. EXTERNAL API ####################
   deploy-rdp-api:
     needs: deployment-environment
-    name: RDP API
+    name: EXTERNAL API
     runs-on: [self-hosted, linux, deployment]
     environment:
       name: ${{ needs.deployment-environment.outputs.environment }}
