+++ .github/workflows/staging_infrastructure.yml
@@ -1062,12 +1062,12 @@ jobs:
         # NB: group-id isn't well documented. It means "Target sub-resource".
         #     The purpose of connection-name is unknown.
         inlineScript: |
-          reference_data_api_id=$(az webapp show --name tfs-${{ env.environment }}-external-api --query "id" -o tsv)
+          external_api_id=$(az webapp show --name tfs-${{ env.environment }}-external-api --query "id" -o tsv)
           az network private-endpoint create \
             --name tfs-${{ env.environment }}-external-api \
             --vnet-name tfs-${{ env.app_service_plan }}-vnet \
             --subnet ${{ env.environment }}-private-endpoints \
-            --private-connection-resource-id "$reference_data_api_id" \
+            --private-connection-resource-id "$external_api_id" \
             --group-id sites \
             --connection-name tfs-${{ env.environment }}-external-api
