+++ azure-functions/number-generator-function/controllers/number-generator.controller.js
@@ -54,4 +54,5 @@ const callNumberGenerator = async (numberType) => {
     };
   }
 };
+
 exports.callNumberGenerator = callNumberGenerator;
