+++ e2e-tests/trade-finance-manager/cypress/support/external-api/api.js
