+++ e2e-tests/trade-finance-manager/cypress/support/portal-api/createFacilities.js
@@ -1,5 +1,5 @@
 const { logIn, createFacilities } = require('./api');
-const { getIdFromNumberGenerator } = require('../reference-data-api/api');
+const { getIdFromNumberGenerator } = require('../external-api/api');
 
 module.exports = (dealId, facilities, user) => {
   console.info('createFacilities::');
