+++ e2e-tests/trade-finance-manager/cypress/support/portal-api/insertManyDeals.js
@@ -1,5 +1,5 @@
 const { insertDeal, getDeal, logIn } = require('./api');
-const { getIdFromNumberGenerator } = require('../reference-data-api/api');
+const { getIdFromNumberGenerator } = require('../external-api/api');
 
 module.exports = (deals, opts) => {
   console.info('createManyDeals::');
