+++ external-api/README.md
@@ -1,12 +1,12 @@
 # EXTERNAL API MICRO-SERVICE 📦️
-RDP is predominately responsible for external API consumption and communication.
+EXTERNAL is predominately responsible for external API consumption and communication.
 ## Endpoints 🌐
 ### 1. APIM
 APIM provides various crucial endpoints across varied services (MDM, TFS and ESTORE).
-RDP has an high exposure to MDM endpoints.
+EXTERNAL has an high exposure to MDM endpoints.
 #### 1. MDM
 Master data management endpoints are responsible for providing wide variety of imperative data.
-They range from `/currencies` to `/interest-rates`, RDP `.env` requires following three variables
+They range from `/currencies` to `/interest-rates`, EXTERNAL `.env` requires following three variables
 to be fulfilled before any consumption.
 
 ```shell
